+++ src/App.css
@@ -18,29 +18,3 @@
   background-color: white;
   color : black;
 }
-
-.table-container{
-  display: block;
-  padding : 20px;
-  width: 90%;
-  border : 1px solid black;
-  background-color: white;
-}
-
-button{
-  margin:  10px 20px 10px 20px;
-}
-
-.hidden {
-  color: white;
-}
-
-.cell{
-  position: relative;
-  background-color: white;
-  padding: 0px 10px
-}
-
-.selected{
-  background: lightskyblue;
-}
