+++ src/components/tableWithModel/TableWithModel.css
@@ -1,6 +1,25 @@
 th td {
     position: relative;
-  }
+}
+
+
+.table-container{
+  display: block;
+  padding : 20px;
+  width: 90%;
+  border : 1px solid black;
+  background-color: white;
+}
+
+.cell{
+  position: relative;
+  background-color: white;
+  padding: 0px 10px
+}
+
+.selected{
+  background: lightskyblue;
+}
 
 .marker{
   border          : 1px solid black;
@@ -14,3 +33,14 @@ th td {
   top : -5px;
   left: 2px;
 }
+
+
+
+button{
+  margin:  10px 20px 10px 20px;
+}
+
+.hidden {
+  color: white;
+}
+
