+++ src/components/tableWithModel/TableWithModel.jsx
@@ -1,13 +1,15 @@
 import React, { useEffect, useState, useRef } from "react";
 import './TableWithModel.css'
-import { getModelFromHtml, isMergedSelected } from "./tableModelUtils";
+import { getModelFromHtml, 
+    // isMergedSelected 
+} from "./tableModelUtils";
 
 
 const TableWithModel = (props) => {
     const [tableModel, setTableModel] = useState( [] )
     const [isCtrlDown, setIsCtrlDown] = useState(false)
 
-
+//** Startup: create TABLE MODEL and add it to DOM */
     useEffect( () => {
         const createModel = () => {
 
@@ -22,14 +24,38 @@ const TableWithModel = (props) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [])
 
+
+
+
+//** *********************************************** */
+//** Ctrl key press handling section */
+
     const ref = useRef(null);
 
     useEffect(() => {
       ref.current.focus();
     }, []);
 
+    const keyDownHandler = (e) => {
+        if ( e.key === 'Control'){
+            setIsCtrlDown( (prev) => { return true })
+        }
+    }
+
+    const keyReleaseHandler = (e) => {
+        if ( e.key === 'Control'){
+            setIsCtrlDown( (prev) => { return false })
+        }
+    }
+
+//** Ctrl key press handling section END */
+//** *********************************************** */
 
 
+
+
+//** *********************************************** */
+//** Click on TABLE handling section */
 
     const clearSelection = () => {
 
@@ -97,40 +123,7 @@ const TableWithModel = (props) => {
         }   
     }
 
-    const pureSelectionListener = ({ target }) => {
-        if( (target.dataset.col === '0') && (target.dataset.row === '0')){
-            clearSelection()
-            props.resetCount()
-            return
-        }
-///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!
-///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!
-///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!
-const headerRowsCount = 4
-///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!
-///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!        
-///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!        
-        
-        const rowNum = target.dataset.row
-        
-        if ( !isCtrlDown ){ 
-            clearSelection()
-        
-            if (rowNum > headerRowsCount - 1){
-                selectRows(rowNum, target.rowSpan)
-            } else {
-                const colNum = Number(target.dataset.col)
-                selectColumns(colNum, target.colSpan)
-            }
-        } else {
-            if (rowNum > headerRowsCount - 1){
-                toggleRows(rowNum, target.rowSpan)
-            } else {
-                const colNum = Number(target.dataset.col)
-                toggleColumns(colNum, target.colSpan)
-            }
-        }
-    }
+
 
 
 
@@ -180,9 +173,71 @@ const headerRowsCount = 4
             toggleRow( toglledRowNum )
         }   
     }
-  
+
+    const pureSelectionListener = ({ target }) => {
+        if( (target.dataset.col === '0') && (target.dataset.row === '0')){
+            clearSelection()
+            props.resetCount()
+            return
+        }
+///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!
+///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!
+///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!
+const headerRowsCount = 4
+///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!
+///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!        
+///REMOVE HARDCODE MAGIC NUMBER !!!!!!!!!!!!!!!!!!        
+        
+        const rowNum = target.dataset.row
+        
+        if ( !isCtrlDown ){ 
+            clearSelection()
+        
+            if (rowNum > headerRowsCount - 1){
+                selectRows(rowNum, target.rowSpan)
+            } else {
+                const colNum = Number(target.dataset.col)
+                selectColumns(colNum, target.colSpan)
+            }
+        } else {
+            if (rowNum > headerRowsCount - 1){
+                toggleRows(rowNum, target.rowSpan)
+            } else {
+                const colNum = Number(target.dataset.col)
+                toggleColumns(colNum, target.colSpan)
+            }
+        }
+    }
+//** Click on TABLE handling section END*/
+//** *********************************************** */
 
 
+
+
+
+//** *********************************************** */
+//** Generate TABLE ELEMENT HTML section  */
+
+
+    const isOneOfMergedSelected = (tableModel, colNum, rowNum, colSpan, rowSpan) => {
+        // console.log('IS SELECTED CHECK : ', colNum, ' ', rowNum, '   ROWSPAN : ', rowSpan)
+        
+        const startCol = colNum
+        const endCol = Number(colNum) + colSpan
+        const startRow = rowNum
+        const endRow = Number(rowNum) + rowSpan
+
+        let result = false
+        
+        for (let col = startCol; col < endCol; col++) {
+            for (let row = startRow; row < endRow; row++) {
+                // console.log('CHECK : X=', col, '   Y=', row) 
+                result = result || tableModel[row][col].selected            
+            }
+        }
+
+        return result
+    }
 
     const Cell = (props) => {
         const cellCol = props.col
@@ -193,12 +248,14 @@ const headerRowsCount = 4
 // selected one of 'back' cells
         let selectedClass
         if ( (cellRowSpan > 1) || (cellColSpan > 1) ){
-            selectedClass = `cell ${isMergedSelected(tableModel, cellCol, cellRow, cellColSpan, cellRowSpan) ? 'selected' : ''}`  
+            selectedClass = `cell ${isOneOfMergedSelected(tableModel, cellCol, cellRow, cellColSpan, cellRowSpan) ? 'selected' : ''}`  
         } else {
             selectedClass = `cell ${tableModel[cellRow][cellCol].selected ? 'selected' : ''}`        
         }
 
-        const marker = (tableModel[cellRow][cellCol].selectCount) ? 
+
+        const marker = 
+            tableModel[cellRow][cellCol].selectCount ? 
             (
                 <div className="marker">
                     {tableModel[cellRow][cellCol].selectCount}
@@ -243,33 +300,10 @@ const headerRowsCount = 4
         )
     }
 
-
-
-    const tableContentHide = () => {
-        const t_container = document.getElementById('table-container')
-        t_container.style.color = t_container.style.color === 'white' ? 'black' : 'white' 
-    }
-
-    const keyDownHandler = (e) => {
-        if ( e.key === 'Control'){
-            setIsCtrlDown( (prev) => { return true})
-        }
-    }
-
-
-    const keyReleaseHandler = (e) => {
-        if ( e.key === 'Control'){
-            setIsCtrlDown( (prev) => { return false})
-        }
-    }
-
-
-
-
-
-    return(
-        <div id="table-container" className="table-container" ref={ref} tabIndex={-1}  onKeyDown={(e)=>keyDownHandler(e)} onKeyUp={(e)=>keyReleaseHandler(e)}> 
-
+    const TableElement = (props) => {
+        const tableModel = props.tableModel
+        
+        return(
             <table border={1}>
                 { tableModel.map( (_r, rowIndex) => {
                     return (
@@ -277,9 +311,44 @@ const headerRowsCount = 4
                     )
                 })}
             </table>
+        )
+    }
+
+//** Generate TABLE ELEMENT HTML section  END*/
+//** *********************************************** */
+
+
+
+
+
+//** *********************************************** */
+//** temporary utils section */
+
+    const tableContentHide = () => {
+        const t_container = document.getElementById('table-container')
+        t_container.style.color = t_container.style.color === 'white' ? 'black' : 'white' 
+    }
+
+    const test  = (e) => {
+        console.log('Test: ', e)
+    }
+
+//** temporary utils section  END*/
+//** *********************************************** */
+
+
+
+
+    return(
+        <div id="table-container" className="table-container" 
+             ref={ref} tabIndex={-1}  
+             onKeyDown={(e)=>keyDownHandler(e)} 
+             onKeyUp={(e)=>keyReleaseHandler(e)}
+            > 
+
+            <TableElement tableModel={tableModel} />
         
-            {/* <button onClick={()=>test()} style={{marginRight: '27px'}}>TEST</button> */}
-            
+            <button onClick={()=>test()} style={{marginRight: '27px'}}>TEST</button>            
             <button onClick={()=>tableContentHide()}>ON | OFF CONTENT</button>
         
         </div>
